backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-dotenv==1.1.1
requests==2.32.5
pydantic==2.11.9
python-multipart==0.0.20
speechrecognition==3.10.0
pydub==0.25.1
pytest==8.3.3
httpx==0.27.2
